+++ Documentation/Changelog.md
@@ -4,6 +4,12 @@
 
 To upgrade from TDW v1.6 to v1.7, read [this guide](Documentation/v1.6_to_v1.7).
 
+## v1.7.7
+
+### Build
+
+- When the build fails to deserialize a list of commands, the error message is more informative.
+
 ## v1.7.6
 
 ### Command API